+++ src/main/java/com/example/profisbackend/dto/studyprogram/StudyProgramResponseDTO.java
@@ -2,5 +2,5 @@ package com.example.profisbackend.dto.studyprogram;
 
 import com.example.profisbackend.enums.DegreeType;
 
-public record StudyProgramResponseDTO(Long id, DegreeType degreeType, String title, float sws) {
+public record StudyProgramResponseDTO(Long id, DegreeType degreeType, String title, double sws) {
 }
